auth-backend/requirements.txt
fastapi>=0.100.0,<0.110.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
uvicorn>=0.23.0,<0.24.0
sqlalchemy>=2.0.0,<3.0.0
passlib>=1.7.4,<1.8.0
python-jose>=3.3.0,<3.4.0
psycopg2-binary>=2.9.1,<2.10.0
alembic>=1.12.0,<1.13.0
python-multipart>=0.0.5,<0.0.6
email-validator>=2.0.0,<3.0.0